Modified

+++ internal/data/quotes.go
@@ -11,6 +11,7 @@ import (
 // represents a quote entry in the sytem
 type Quotes struct {
 	Quote_id   int64     `json:"quote_id"`
+	User_id    int64     `json:"user_id"`
 	Content    string    `json:"content"`
 	Created_at time.Time `json:"created_at"`
 }
@@ -30,8 +31,8 @@ type QuotesModel struct {
 // Adds new todo entry into the database
 func (m *QuotesModel) Insert(quotes *Quotes) error {
 	query := `
-		INSERT INTO quotes (content)
-		VALUES ($1)
+		INSERT INTO quotes (content, user_id)
+		VALUES ($1,$2)
 		RETURNING quote_id, created_at`
 
 	ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
@@ -41,20 +42,22 @@ func (m *QuotesModel) Insert(quotes *Quotes) error {
 		ctx,
 		query,
 		quotes.Content,
+		quotes.User_id,
 	).Scan(&quotes.Quote_id, &quotes.Created_at)
 }
 
 // Retrieve list of all quote entries from the database
-func (m *QuotesModel) QuoteList() ([]*Quotes, error) {
+func (m *QuotesModel) QuoteList(userID int64) ([]*Quotes, error) {
 	query := `
-        SELECT quote_id, content
+        SELECT quote_id, content, user_id, created_at
         FROM quotes
+        WHERE user_id = $1
         ORDER BY created_at DESC`
 
 	ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
 	defer cancel()
 
-	rows, err := m.DB.QueryContext(ctx, query)
+	rows, err := m.DB.QueryContext(ctx, query, userID)
 	if err != nil {
 		return nil, err
 	}
@@ -64,7 +67,7 @@ func (m *QuotesModel) QuoteList() ([]*Quotes, error) {
 
 	for rows.Next() {
 		q := &Quotes{}
-		err := rows.Scan(&q.Quote_id, &q.Content)
+		err := rows.Scan(&q.Quote_id, &q.Content, &q.User_id, &q.Created_at)
 		if err != nil {
 			return nil, err
 		}
@@ -78,14 +81,26 @@ func (m *QuotesModel) QuoteList() ([]*Quotes, error) {
 	return quotes, nil
 }
 
-// DeleteQuote removes a quote entry from the database using its ID
-func (m *QuotesModel) DeleteQuote(quoteID int64) error {
+func (m *QuotesModel) DeleteQuote(quoteID int64, userID int64) error {
 	query := `
-	DELETE FROM quotes WHERE quote_id = $1`
+    DELETE FROM quotes WHERE quote_id = $1 AND user_id = $2`
 
 	ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
 	defer cancel()
 
-	_, err := m.DB.ExecContext(ctx, query, quoteID)
-	return err
+	result, err := m.DB.ExecContext(ctx, query, quoteID, userID)
+	if err != nil {
+		return err
+	}
+
+	rowsAffected, err := result.RowsAffected()
+	if err != nil {
+		return err
+	}
+
+	if rowsAffected == 0 {
+		return sql.ErrNoRows
+	}
+
+	return nil
 }
